Fix problem with webviews spilling out beyond notebook and consolidate logic a bit better.
