+++ src/vs/workbench/contrib/positronNotebook/browser/PositronNotebookComponent.tsx
@@ -32,7 +32,9 @@ export function PositronNotebookComponent() {
 	const fontStyles = useFontStyles();
 	const containerRef = React.useRef<HTMLDivElement>(null);
 
-	useScrollContainerEvents(containerRef);
+	React.useEffect(() => {
+		notebookInstance.setCellsContainer(containerRef.current);
+	}, [notebookInstance]);
 
 	return (
 		<div className='positron-notebook' style={{ ...fontStyles }}>
@@ -63,53 +65,6 @@ function useFontStyles(): React.CSSProperties {
 		'--vscode-positronNotebook-text-output-font-family': family,
 		'--vscode-positronNotebook-text-output-font-size': `${fontInfo.fontSize}px`,
 	} as React.CSSProperties;
-}
-
-/**
- * Hook to manage scroll and DOM mutation events for the notebook cells container
- */
-function useScrollContainerEvents(
-	containerRef: React.RefObject<HTMLDivElement>,
-) {
-	const notebookInstance = useNotebookInstance();
-
-	React.useEffect(() => {
-		const container = containerRef.current;
-		if (!container) {
-			return;
-		}
-
-		// Fire initial scroll event after a small delay to ensure layout has settled
-		const initialScrollTimeout = setTimeout(() => {
-			notebookInstance.fireOnDidScrollCellsContainer();
-		}, 50);
-
-		// Set up scroll listener
-		const scrollListener = DOM.addDisposableListener(container, 'scroll', () => {
-			notebookInstance.fireOnDidScrollCellsContainer();
-		});
-
-		// Set up mutation observer to watch for DOM changes
-		const observer = new MutationObserver(() => {
-			// Small delay to let the DOM changes settle
-			setTimeout(() => {
-				notebookInstance.fireOnDidScrollCellsContainer();
-			}, 0);
-		});
-
-		observer.observe(container, {
-			childList: true,
-			subtree: true,
-			attributes: true,
-			attributeFilter: ['style', 'class']
-		});
-
-		return () => {
-			clearTimeout(initialScrollTimeout);
-			scrollListener.dispose();
-			observer.disconnect();
-		};
-	}, [notebookInstance]);
 }
 
 function NotebookCell({ cell }: {
