+++ src/vs/workbench/contrib/positronNotebook/browser/notebookCells/PreloadMessageOutput.tsx
@@ -42,7 +42,7 @@ export function PreloadMessageOutput({ preloadMessageResult }: PreloadMessageOut
 }
 
 function DisplayedPreloadMessage({ webview }: DisplayedPreloadMessageProps) {
-	const { containerRef, clipContainerRef, isLoading, error } = useWebviewMount(webview);
+	const { containerRef, isLoading, error } = useWebviewMount(webview);
 
 	if (error) {
 		return <div>{error.message}</div>;
@@ -51,13 +51,7 @@ function DisplayedPreloadMessage({ webview }: DisplayedPreloadMessageProps) {
 	return (
 		<>
 			{isLoading && <div>{MESSAGES.LOADING}</div>}
-			<div
-				ref={clipContainerRef}
-				className='positron-webview-output-clip-container'
-				style={{ position: 'relative' }}
-			>
-				<div style={{ position: 'absolute', top: 0, left: 0, width: '100%', height: '100%' }} ref={containerRef} />
-			</div>
+			<div ref={containerRef} />
 		</>
 	);
 }
