+++ src/vs/workbench/services/positronNotebook/browser/IPositronNotebookInstance.ts
@@ -57,6 +57,18 @@ export interface IPositronNotebookInstance {
 	readonly connectedToEditor: boolean;
 
 	/**
+	 * The DOM element that contains the cells for the notebook.
+	 * This is set when the cells container is mounted in the React component.
+	 */
+	readonly cellsContainer: HTMLElement | undefined;
+
+	/**
+	 * Sets the DOM element that contains the cells for the notebook.
+	 * @param container The container element to set, or undefined to clear
+	 */
+	setCellsContainer(container: HTMLElement | undefined): void;
+
+	/**
 	 * Observable array of cells that make up the notebook. Changes to this array
 	 * will trigger UI updates in connected views.
 	 */
